Rename TokensApiProvider types
