+++ packages/providers/src/ports/TokensApiProvider/index.ts
@@ -5,4 +5,4 @@ export {
   MintCodeInput,
   Transaction,
   TransactionStatus,
-} from './Types';
+} from './types';
