+++ packages/providers/src/ports/TokensApiProvider/types/MintCodeResponse.ts
@@ -52,15 +52,3 @@ interface Drop {
   event_template_id?: number | null;
   private_event: boolean;
 }
-
-export enum TransactionStatus {
-  pending = 'pending',
-  passed = 'passed',
-  failed = 'failed',
-  waiting = 'waiting_tx',
-}
-
-export interface Transaction {
-  tx_hash: string;
-  status: TransactionStatus;
-}
