+++ packages/providers/src/ports/TokensApiProvider/types/Transaction.ts
@@ -1,0 +1,11 @@
+export enum TransactionStatus {
+  pending = 'pending',
+  passed = 'passed',
+  failed = 'failed',
+  waiting = 'waiting_tx',
+}
+
+export interface Transaction {
+  tx_hash: string;
+  status: TransactionStatus;
+}
